modbusmup/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
click==8.1.3
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
django-rq==2.6.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.1
redis==4.5.4
rq==1.12.0
sqlparse==0.4.2
